Rebases and reworks illustrated message
Cleans up code

Fixes changes to unrelated files that leaked un during rebase

Fixes comment typo
