+++ src/button/spectrum-action-button.css
@@ -26,26 +26,14 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
     align-items: center;
     -ms-flex-pack: center;
     justify-content: center;
-
-    /* Show the button overflow in Edge. */
     overflow: visible;
-
-    /* Remove button the margin in Firefox and Safari. */
     margin: 0;
 
     border-style: solid;
-    white-space: nowrap;
-
-    /* Remove the inheritance of text transform on button in Edge, Firefox, and IE. */
     text-transform: none;
     -webkit-font-smoothing: antialiased;
-    /* Font smoothing for Firefox */
     -moz-osx-font-smoothing: grayscale;
-
-    /* Correct the inability to style clickable types in iOS and Safari. */
     -webkit-appearance: button;
-
-    /* Adjacent buttons should be aligned correctly */
     vertical-align: top;
 
     transition: background var(--spectrum-global-animation-duration-100, 130ms)
@@ -58,6 +46,8 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
     text-decoration: none;
     font-family: adobe-clean, 'Source Sans Pro', -apple-system,
         BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+
+    line-height: 1.3;
 
     cursor: pointer;
 
@@ -73,8 +63,6 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
         --spectrum-actionbutton-min-width,
         var(--spectrum-global-dimension-size-400)
     );
-
-    /* Use icon padding by default as it's smaller */
     padding: 0
         calc(
             var(
@@ -146,12 +134,8 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
    * .spectrum-ClearButton::-moz-focus-inner,
    * .spectrum-Tool::-moz-focus-inner */
     border: 0;
-
-    /* Remove the inner border and padding for button in Firefox. */
     border-style: none;
     padding: 0;
-
-    /* Use uppercase PX so values don't get converted to rem */
     margin-top: -2px;
     margin-bottom: -2px;
 }
@@ -208,13 +192,10 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
 .spectrum-Icon + #label {
     /* .spectrum-ActionButton .spectrum-Icon + .spectrum-ActionButton-label,
    * .spectrum-Tool .spectrum-Icon + .spectrum-ActionButton-label */
-    /* Have icon padding on the left */
     padding-left: var(
         --spectrum-actionbutton-icon-padding-x,
         var(--spectrum-global-dimension-size-85)
     );
-
-    /* Have text padding on the right */
     padding-right: calc(
         var(
                 --spectrum-actionbutton-text-padding-x,
@@ -229,7 +210,6 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
 .spectrum-Icon--sizeS:only-child {
     /* .spectrum-ActionButton .spectrum-Icon--sizeS:only-child,
    * .spectrum-Tool .spectrum-Icon--sizeS:only-child */
-    /* Position absolutely to avoid layout errors introduced by padding */
     position: absolute;
     top: calc(
         50% -
@@ -249,7 +229,6 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
 #label:only-child {
     /* .spectrum-ActionButton .spectrum-ActionButton-label:only-child,
    * .spectrum-Tool .spectrum-ActionButton-label:only-child */
-    /* Add padding for text only buttons */
     padding: 0
         calc(
             var(
@@ -283,15 +262,15 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
 #label {
     /* .spectrum-ActionButton-label,
    * .spectrum-Button-label */
-    display: -ms-flexbox;
-    display: flex;
-    -ms-flex-pack: center;
-    justify-content: center;
-    -ms-flex-align: center;
-    align-items: center;
-
-    /* @safari10 Workaround for https://bugs.webkit.org/show_bug.cgi?id=169700 */
+    -ms-flex-item-align: center;
+    -ms-grid-row-align: center;
+    align-self: center;
+    justify-self: center;
     width: 100%;
+    /* .spectrum-ActionButton-label */
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
 }
 #label:empty {
     /* .spectrum-ActionButton-label:empty,
