+++ src/button/spectrum-button.css
@@ -26,26 +26,14 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
     align-items: center;
     -ms-flex-pack: center;
     justify-content: center;
-
-    /* Show the button overflow in Edge. */
     overflow: visible;
-
-    /* Remove button the margin in Firefox and Safari. */
     margin: 0;
 
     border-style: solid;
-    white-space: nowrap;
-
-    /* Remove the inheritance of text transform on button in Edge, Firefox, and IE. */
     text-transform: none;
     -webkit-font-smoothing: antialiased;
-    /* Font smoothing for Firefox */
     -moz-osx-font-smoothing: grayscale;
-
-    /* Correct the inability to style clickable types in iOS and Safari. */
     -webkit-appearance: button;
-
-    /* Adjacent buttons should be aligned correctly */
     vertical-align: top;
 
     transition: background var(--spectrum-global-animation-duration-100, 130ms)
@@ -59,6 +47,8 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
     font-family: adobe-clean, 'Source Sans Pro', -apple-system,
         BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
 
+    line-height: 1.3;
+
     cursor: pointer;
 
     /* .spectrum-Button */
@@ -70,32 +60,31 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
         var(--spectrum-global-dimension-size-200)
     );
 
-    height: var(
+    min-height: var(
         --spectrum-button-primary-height,
         var(--spectrum-global-dimension-size-400)
     );
+    height: auto;
     min-width: var(
         --spectrum-button-primary-min-width,
         var(--spectrum-button-cta-min-width)
     );
 
-    padding: 0
+    padding: var(--spectrum-global-dimension-size-50)
         calc(
             var(
                     --spectrum-button-primary-padding-x,
                     var(--spectrum-global-dimension-size-200)
                 ) - var(--spectrum-button-primary-border-size, 2px)
-        )
-        0;
+        );
+    padding-bottom: calc(var(--spectrum-global-dimension-size-50) + 0.5px);
+    padding-top: calc(var(--spectrum-global-dimension-size-50) - 0.5px);
 
     font-size: var(
         --spectrum-button-primary-text-size,
         var(--spectrum-global-dimension-font-size-150)
     );
     font-weight: var(--spectrum-button-primary-text-font-weight, 700);
-
-    /* Prevent vertical text alignment mismatch between anchor and button in Chrome */
-    line-height: 0;
 }
 #button:focus {
     /* .spectrum-Button:focus,
@@ -118,12 +107,8 @@ THIS FILE IS MACHINE GENERATED. DO NOT EDIT */
    * .spectrum-ClearButton::-moz-focus-inner,
    * .spectrum-Tool::-moz-focus-inner */
     border: 0;
-
-    /* Remove the inner border and padding for button in Firefox. */
     border-style: none;
     padding: 0;
-
-    /* Use uppercase PX so values don't get converted to rem */
     margin-top: -2px;
     margin-bottom: -2px;
 }
@@ -181,14 +166,10 @@ slot[name='icon'] + #label {
 #label {
     /* .spectrum-ActionButton-label,
    * .spectrum-Button-label */
-    display: -ms-flexbox;
-    display: flex;
-    -ms-flex-pack: center;
-    justify-content: center;
-    -ms-flex-align: center;
-    align-items: center;
-
-    /* @safari10 Workaround for https://bugs.webkit.org/show_bug.cgi?id=169700 */
+    -ms-flex-item-align: center;
+    -ms-grid-row-align: center;
+    align-self: center;
+    justify-self: center;
     width: 100%;
 }
 #label:empty {
@@ -198,7 +179,6 @@ slot[name='icon'] + #label {
 }
 #button:active {
     /* .spectrum-Button:active */
-    /* Override focus -- clicking with spacebar should not show outline */
     box-shadow: none;
 }
 :host([variant='cta']) #button {
