--- src/illustrated-message/test/index.test.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <title>Spectrum Web Components - Illustrated Message Tests</title>
-        <script src="/node_modules/karma-web-components/framework.js"></script>
-        <link rel="stylesheet" href="../../../styles/all-medium-light.css" />
-    </head>
-
-    <body>
-        <sp-illustrated-message
-            heading="No Results"
-            description="Try another search"
-        >
-            <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="150"
-                height="103"
-                viewBox="0 0 150 103"
-            >
-                <use xlink:href="error_message_geometry.svg#error-notice" />
-            </svg>
-        </sp-illustrated-message>
-        <script src="./index.js" type="module"></script>
-    </body>
-</html>
